Moved the layout folder to the app folder
- These layout will be used in the app, so they should be in the main folder
