+++ src/app/layouts/blank/blank.component.ts
