+++ src/app/layouts/full/horizontal/header/search-dialog.component.html
