+++ src/app/layouts/full/shared/customizer/customizer.component.scss
