+++ src/app/layouts/full/vertical/header/header.component.ts
