+++ src/app/layouts/full/vertical/header/search-dialog.component.html
