+++ src/app/layouts/full/vertical/sidebar/branding.component.ts
