+++ src/app/layouts/full/vertical/sidebar/nav-item/nav-item.component.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core';
 import { NavItem } from './nav-item';
 import { Router } from '@angular/router';
-import { NavService } from '../../../../../../services/nav.service';
+import { NavService } from '../../../../../services/nav.service';
 import {
   animate,
   state,
