+++ src/app/layouts/full/vertical/sidebar/sidebar-data.ts
