+++ src/app/layouts/full/vertical/sidebar/sidebar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { NavService } from '../../../../../services/nav.service';
+import { NavService } from '../../../../services/nav.service';
 import { NgScrollbarModule } from 'ngx-scrollbar';
 import { TablerIconsModule } from 'angular-tabler-icons';
 import { MaterialModule } from 'src/app/material.module';
